refactoring to use more raster data and not loop over pixels

+++ src/main/scala/net/retorx/glitchvideo/glitches/colorbands/NoiseLinesSingleColorBands.scala
@@ -21,7 +21,7 @@ class NoiseLinesSingleColorBands(lineThickness: Int = 1, colorBand: Color = RED)
         }
 
         if (line >= y && line <= (y + thickness)) {
-            colorBand.partiallyFuck(r, g, b)
+            colorBand.partiallyFuck(r.toByte, g.toByte, b.toByte)
         } else {
             r | g | b
         }
